+++ thoipapy/experimental_data/ttest_features.py
@@ -47,13 +47,13 @@ def conduct_ttest_for_all_features(s, logging):
     logging.info('starting conduct_ttest_for_selected_features_used_in_model')
     testsetname, trainsetname = get_testsetname_trainsetname_from_run_settings(s)
     # inputs
-    train_data_csv = Path(s["thoipapy_data_folder"]) / f"results/{s['setname']}/train_data/01_train_data_orig.csv"
+    train_data_csv = Path(s["data_dir"]) / f"results/{s['setname']}/train_data/01_train_data_orig.csv"
     # IMPORTANT: the features used in the model are taken from the trainset as defined in "train_datasets" in the excel file, not from the actual set being investigated
-    # feat_imp_MDA_xlsx = os.path.join(s["thoipapy_data_folder"], "results", trainsetname, "feat_imp", "feat_imp_mean_decrease_accuracy.xlsx")
+    # feat_imp_MDA_xlsx = os.path.join(s["data_dir"], "results", trainsetname, "feat_imp", "feat_imp_mean_decrease_accuracy.xlsx")
 
     # outputs
-    ttest_pvalues_bootstrapped_data_using_traindata_selected_features_xlsx = Path(s["thoipapy_data_folder"]) / f"results/{s['setname']}/ttest/ttest_pvalues_bootstrapped_data_using_traindata_selected_features(train{trainsetname}).xlsx"
-    correlated_features_xlsx = Path(s["thoipapy_data_folder"]) / f"results/{s['setname']}/ttest/correlated_features.xlsx"
+    ttest_pvalues_bootstrapped_data_using_traindata_selected_features_xlsx = Path(s["data_dir"]) / f"results/{s['setname']}/ttest/ttest_pvalues_bootstrapped_data_using_traindata_selected_features(train{trainsetname}).xlsx"
+    correlated_features_xlsx = Path(s["data_dir"]) / f"results/{s['setname']}/ttest/correlated_features.xlsx"
 
     make_sure_path_exists(ttest_pvalues_bootstrapped_data_using_traindata_selected_features_xlsx, isfile=True)
 
@@ -62,7 +62,7 @@ def conduct_ttest_for_all_features(s, logging):
     # df_feat_imp_MDA_trainset = pd.read_excel(feat_imp_MDA_xlsx, sheet_name="single_feat", index_col=0)
     # cols = list(df_feat_imp_MDA_trainset.index)
 
-    df = drop_cols_not_used_in_ML(logging, df_orig, s["excel_file_with_settings"])
+    df = drop_cols_not_used_in_ML(logging, df_orig, s["settings_path"])
     feature_columns = list(df.columns)
 
     # add back the interface "bind" column
@@ -107,8 +107,8 @@ def conduct_ttest_for_all_features(s, logging):
 
     for col in feature_columns:
         correlated_features = dfcorr[col].loc[dfcorr[col] > cutoff_R2_correlated].index.tolist()
-        correlated_features.remove(col)
         if len(correlated_features) > 0:
+            correlated_features.remove(col)
             dft.loc[col, "correlated features (R2 > 0.6)"] = ", ".join(correlated_features)
         else:
             dft.loc[col, "correlated features (R2 > 0.6)"] = ""