+++ thoipapy/features/lips.py
@@ -28,15 +28,15 @@ def LIPS_score_calculation_mult_prot(s, df_set, logging):
     for i in df_set.index:
         acc = df_set.loc[i, "acc"]
         database = df_set.loc[i, "database"]
-        # LIPS_input_file = os.path.join(s["thoipapy_data_folder"], "homologues", "a3m",database, "%s.mem.lips.input%s") % (acc,s["surres"])
-        alignments_dir = os.path.join(s["thoipapy_data_folder"], "homologues", "alignments", database)
+        # LIPS_input_file = os.path.join(s["data_dir"], "homologues", "a3m",database, "%s.mem.lips.input%s") % (acc,s["surres"])
+        alignments_dir = os.path.join(s["data_dir"], "homologues", "alignments", database)
         path_uniq_TMD_seqs_no_gaps_for_LIPS = os.path.join(alignments_dir, "{}.surr{}.gaps0.uniq.for_LIPS.txt".format(acc, s["num_of_sur_residues"]))
 
         if os.path.isfile(path_uniq_TMD_seqs_no_gaps_for_LIPS):
-            # LIPS_output_file = os.path.join(s["thoipapy_data_folder"], "features", "lips_score", "zpro/NoRedundPro/%s.mem.lips.output") % acc
+            # LIPS_output_file = os.path.join(s["data_dir"], "features", "lips_score", "zpro/NoRedundPro/%s.mem.lips.output") % acc
             # path_uniq_TMD_seqs_no_gaps_for_LIPS = os.path.join(alignments_dir, "{}.surr{}.gaps0.uniq.for_LIPS.txt".format(acc, s["num_of_sur_residues"]))
 
-            # LIPS_output_file = os.path.join(s["thoipapy_data_folder"], "features", "lips_score", database, "%s.mem.lips.output%s") % (acc, s["surres"])
+            # LIPS_output_file = os.path.join(s["data_dir"], "features", "lips_score", database, "%s.mem.lips.output%s") % (acc, s["surres"])
 
             LIPS_output_file = os.path.join(alignments_dir, "{}.surr{}.LIPS_output.csv".format(acc, s["num_of_sur_residues"]))
 
@@ -323,9 +323,9 @@ def parse_LIPS_score_mult_prot(s, df_set, logging):
     for i in df_set.index:
         acc = df_set.loc[i, "acc"]
         database = df_set.loc[i, "database"]
-        alignments_dir = os.path.join(s["thoipapy_data_folder"], "homologues", "alignments", database)
+        alignments_dir = os.path.join(s["data_dir"], "homologues", "alignments", database)
         LIPS_output_file = os.path.join(alignments_dir, "{}.surr{}.LIPS_output.csv".format(acc, s["num_of_sur_residues"]))
-        LIPS_parsed_csv = os.path.join(s["thoipapy_data_folder"], "features", "lips_score", database, "{}.surr{}.LIPS_score_parsed.csv".format(acc, s["num_of_sur_residues"]))
+        LIPS_parsed_csv = os.path.join(s["data_dir"], "features", "lips_score", database, "{}.surr{}.LIPS_score_parsed.csv".format(acc, s["num_of_sur_residues"]))
         parse_LIPS_score(acc, LIPS_output_file, LIPS_parsed_csv, logging)