+++ thoipapy/other/other_figs/average_fraction_DI.py
@@ -64,7 +64,7 @@ def FigZB_07_hardlinked(Fontsize, Width, Size, s):
 
     # read name from name file
     #name_file = r"H:\data_xy\protein_names.xlsx"
-    names_excel_path = os.path.join(s["dropbox_dir"], "ETRA_NMR_names.xlsx")
+    names_excel_path = os.path.join(s["base_dir"], "ETRA_NMR_names.xlsx")
     df_name_file = pd.read_excel(names_excel_path, sheet_name='proteins', index_col=0)